pyproject.toml
[project]
name = "odsc-agentic-ai-summit-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "baml-py==0.203.1",
    "kuzu>=0.11.0",
    "lancedb>=0.24.0",
    "ollama>=0.5.1",
    "polars>=1.31.0",
    "pandas>=2.3.1",
    "pyarrow>=20.0.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "fastapi>=0.116.0",
    "uvicorn>=0.35.0",
    "httpx>=0.28.1",
    "opik>=1.8.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]
